StudyTest/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
configparser==5.2.0
ddt==1.6.0
exceptiongroup==1.0.4
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
multi-key-dict==2.0.3
packaging==21.3
pbr==5.9.0
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
python-jenkins==1.7.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
